Refactor image funcs from vision to image.py
